==> notebooks/3. curated/4. data_joining_pt3.ipynb <==
import pandas as pd
import sys
import os

sys.path.append(os.path.abspath('../../scripts/2. modules'))
import processing

# Inputs
JOINED_PATH_LEFT = '../../data/3. curated/merged/crime_schools_distances_land.csv'
JOINED_PATH_RIGHT = '../../data/3. curated/merged/external_prices.csv'

# Outputs
OUTPUT_PATH = '../../data/3. curated/merged'
OUTPUT_NAME = 'all_historical_data.csv'

df_left = pd.read_csv(JOINED_PATH_LEFT)
df_right = pd.read_csv(JOINED_PATH_RIGHT)

df_joined = pd.merge(df_left, df_right, on=['suburbs', 'year', 'quarter'])

Calculate per capita for the numeric features

merged_df_per_capita = df_joined.copy()

PER_CAPITA_COLUMNS = [ 'a20_assault_and_related_offences',
 'a50_robbery',
 'a70_stalking,_harassment_and_threatening_behaviour',
 'a80_dangerous_and_negligent_acts_endangering_people',
 'other_crimes_against_the_person',
 'b10_arson',
 'b20_property_damage',
 'b30_burglary/break_and_enter',
 'b40_theft',
 'b50_deception',
 'c10_drug_dealing_and_trafficking',
 'c20_cultivate_or_manufacture_drugs',
 'c30_drug_use_and_possession',
 'd10_weapons_and_explosives_offences',
 'd20_disorderly_and_offensive_conduct',
 'd30_public_nuisance_offences',
 'e10_justice_procedures',
 'e20_breaches_of_orders',
 'f20_transport_regulation_offences',
 'f90_miscellaneous_offences',
 'c90_other_drug_offences',
 'f30_other_government_regulatory_offences',
 'd40_public_security_offences',
 'f10_regulatory_driving_offences',
 'b60_bribery',
 'total_crimes',
 'economic: median age of earners',
 'economic: median income',
 'economic: top 5% suburb',
 'economic: top 1% suburb',
 'economic: gini coefficient',
 'economic: trimmed mean quarterly',
 'economic: variable interest rate',
 'economic: gdp quarterly',
 'economic: gdp per capita quarterly',
 'population: 0-4',
 'population: 5-9',
 'population: 10-14',
 'population: 15-19',
 'population: 20-24',
 'population: 25-29',
 'population: 30-34',
 'population: 35-39',
 'population: 40-44',
 'population: 45-49',
 'population: 50-54',
 'population: 55-59',
 'population: 60-64',
 'population: 65-69',
 'population: 70-74',
 'population: 75-79',
 'population: 80-84',
 'population: 85 and over',
 'birth: brazil',
 'birth: wales',
 'birth: papua new guinea',
 'birth: australia',
 'birth: malaysia',
 'birth: country of birth not stated',
 'birth: indonesia',
 'birth: egypt',
 'birth: afghanistan age',
 'birth: thailand',
 'birth: ireland',
 'birth: hong kong sar of china',
 'birth: nepal',
 'birth: north macedonia',
 'birth: england',
 'birth: chile',
 'birth: fiji',
 'birth: taiwan',
 'birth: japan',
 'birth: malta',
 'birth: scotland',
 'birth: china excludes sars and taiwan',
 'birth: poland',
 'birth: singapore',
 'birth: south africa',
 'birth: greece',
 'birth: pakistan',
 'birth: iraq',
 'birth: mauritius',
 'birth: samoa',
 'birth: sri lanka',
 'birth: india',
 'birth: philippines',
 'birth: vietnam',
 'birth: new zealand',
 'birth: iran',
 'birth: lebanon',
 'birth: france',
 'birth: total',
 'birth: netherlands',
 'birth: myanmar',
 'birth: united states of america',
 'birth: bosnia and herzegovina',
 'birth: zimbabwe',
 'birth: bangladesh',
 'birth: turkey',
 'birth: canada',
 'birth: korea republic of south',
 'birth: cambodia',
 'birth: born elsewhere',
 'birth: italy',
 'birth: croatia',
 'birth: germany',
 'studying: primary other',
 'studying: tertiary FT 14-25',
 'studying: tertiary FT 25+',
 'studying: primary total',
 'studying: tertiary PT 14-25',
 'studying: tertiary PT 25+',
 'studying: tafe total',
 'studying: primary government',
 'studying: secondary catholic',
 'studying: preschool',
 'studying: primary catholic',
 'studying: secondary government',
 'studying: secondary total',
 'studying: tertiary total',
 'overseas: 5 years',
 'relationships: child under 15',
 'relationships: defacto',
 'relationships: other related individual',
 'relationships: non dependent child',
 'relationships: lone persons',
 'relationships: group household',
 'relationships: married',
 'relationships: dependent student',
 'relationships: lone parents',
 'economic: number of earners',
 ]

for col in PER_CAPITA_COLUMNS:
    merged_df_per_capita[col + '_pc'] = merged_df_per_capita[col] / merged_df_per_capita['population: total persons']

merged_df_per_capita = merged_df_per_capita.copy() # .drop(['Unnamed: 0.1', 'Unnamed: 0',], axis=1)

processing.to_csv(merged_df_per_capita, OUTPUT_PATH, OUTPUT_NAME)
